+++ common/systems/src/beam.rs
@@ -2,6 +2,7 @@ use common::{
     combat::{self, AttackOptions, AttackSource, AttackerInfo, TargetInfo},
     comp::{
         agent::{Sound, SoundKind},
+        aura::EnteredAuras,
         Alignment, Beam, Body, Buffs, CharacterState, Combo, Energy, Group, Health, Inventory, Ori,
         Player, Pos, Scale, Stats,
     },
@@ -59,6 +60,7 @@ pub struct ReadData<'a> {
     combos: ReadStorage<'a, Combo>,
     character_states: ReadStorage<'a, CharacterState>,
     buffs: ReadStorage<'a, Buffs>,
+    entered_auras: ReadStorage<'a, EnteredAuras>,
     outcomes: Read<'a, EventBus<Outcome>>,
     events: ReadAttackEvents<'a>,
 }
@@ -203,6 +205,12 @@ impl<'a> System<'a> for Sys {
                                 >= tgt_dist;
 
                         if hit {
+                            let allow_friendly_fire = combat::allow_friendly_fire(
+                                &read_data.entered_auras,
+                                entity,
+                                target,
+                            );
+
                             // See if entities are in the same group
                             let same_group = group
                                 .map(|group_a| Some(group_a) == read_data.groups.get(target))
@@ -243,9 +251,10 @@ impl<'a> System<'a> for Sys {
                                 .and_then(|cs| cs.attack_immunities())
                                 .map_or(false, |i| i.beams);
                             // PvP check
-                            let may_harm = combat::may_harm(
+                            let permit_pvp = combat::permit_pvp(
                                 &read_data.alignments,
                                 &read_data.players,
+                                &read_data.entered_auras,
                                 &read_data.id_maps,
                                 Some(entity),
                                 target,
@@ -275,7 +284,8 @@ impl<'a> System<'a> for Sys {
 
                             let attack_options = AttackOptions {
                                 target_dodging,
-                                may_harm,
+                                permit_pvp,
+                                allow_friendly_fire,
                                 target_group,
                                 precision_mult,
                             };