+++ common/systems/src/melee.rs
@@ -2,6 +2,7 @@ use common::{
     combat::{self, AttackOptions, AttackSource, AttackerInfo, TargetInfo},
     comp::{
         agent::{Sound, SoundKind},
+        aura::EnteredAuras,
         melee::MultiTarget,
         Alignment, Body, Buffs, CharacterState, Combo, Energy, Group, Health, Inventory, Melee,
         Ori, Player, Pos, Scale, Stats,
@@ -59,6 +60,7 @@ pub struct ReadData<'a> {
     stats: ReadStorage<'a, Stats>,
     combos: ReadStorage<'a, Combo>,
     buffs: ReadStorage<'a, Buffs>,
+    entered_auras: ReadStorage<'a, EnteredAuras>,
     events: ReadAttackEvents<'a>,
 }
 
@@ -186,6 +188,8 @@ impl<'a> System<'a> for Sys {
                     && !is_blocked_by_wall(&read_data.terrain, attacker_cylinder, target_cylinder);
 
                 if hit {
+                    let allow_friendly_fire =
+                        combat::allow_friendly_fire(&read_data.entered_auras, attacker, target);
                     // See if entities are in the same group
                     let same_group = read_data
                         .groups
@@ -227,9 +231,10 @@ impl<'a> System<'a> for Sys {
                     };
 
                     // PvP check
-                    let may_harm = combat::may_harm(
+                    let permit_pvp = combat::permit_pvp(
                         &read_data.alignments,
                         &read_data.players,
+                        &read_data.entered_auras,
                         &read_data.id_maps,
                         Some(attacker),
                         target,
@@ -264,7 +269,8 @@ impl<'a> System<'a> for Sys {
 
                     let attack_options = AttackOptions {
                         target_dodging,
-                        may_harm,
+                        permit_pvp,
+                        allow_friendly_fire,
                         target_group,
                         precision_mult,
                     };