+++ common/systems/tests/phys/utils.rs
@@ -1,7 +1,8 @@
 use common::{
     comp::{
         inventory::item::MaterialStatManifest, tool::AbilityMap, Auras, Buffs, CharacterActivity,
-        CharacterState, Collider, Combo, Controller, Energy, Health, Ori, Pos, Stats, Vel,
+        CharacterState, Collider, Combo, Controller, Energy, EnteredAuras, Health, Ori, Pos, Stats,
+        Vel,
     },
     resources::{DeltaTime, GameMode, Time},
     shared_server_config::ServerConstants,
@@ -128,6 +129,7 @@ pub fn create_player(state: &mut State) -> Entity {
         .with(Buffs::default())
         .with(Combo::default())
         .with(Auras::default())
+        .with(EnteredAuras::default())
         .with(Energy::new(body))
         .with(Health::new(body))
         .with(skill_set)